Got tests passing
